tank_id implemented

+++ src/game/GameObject.cpp
@@ -4,6 +4,6 @@
 GameObject::GameObject(std::vector<Position> positions, const Position &translation, float rotation,
                        const Position &rotationPoint) : positions(std::move(positions)), translation(translation),
                                                         rotation(rotation), rotation_point(rotationPoint) {
-    static unsigned long long counter = 0;
+    static unsigned int counter = 0;
     id = counter++;
 }
